Normalize type names when creating Ruby classes
`GraphQL::Client::Schema.generate` creates a hierarchy of Ruby modules
and classes based on a GraphQL schema, but Ruby requires constants to
start with a capital letter, which causes problems when types in the
GraphQL schema start with a lowercase letter, as they're allowed to do.

The solution is to convert any lowercase type names into UpperCamelCase
before creating their associated constants. Type names that start with a
capital letter are left alone, so this change shouldn't affect any
schemas that were capable of being parsed by the existing code.

Fixes #123