Use safe_constantize instead of rescuing
The rescue that was here previously can be over-eager in capturing
exceptions, and could mask error that weren't specifically that
const_name didn't exist but could be any NameError that happened eager
loading that class or that class's dependencies.

For this reason ActiveSupport provides safe_constantize which returns
nil only if the NameError comes from the constant we are requesting, and
raises the exception otherwise.